requirements.txt
astroid==1.6.1
autopep8==1.3.5
boto3==1.6.6
botocore==1.9.6
certifi==2018.1.18
chardet==3.0.4
cloudinary==1.11.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.10
django-crispy-forms==1.7.0
django-markdown-deux==1.0.5
django-pagedown==0.1.3
-e git+https://github.com/gbozee/django-paystack.git@0a64fa970a53799b33adef8984c3b9eb1d8b37aa#egg=django_paystack
django-storages==1.6.5
django-toolbelt==0.0.1
docopt==0.4.0
docutils==0.14
future==0.16.0
gunicorn==19.7.1
idna==2.6
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mailchimp==2.0.9
markdown2==2.3.5
mccabe==0.6.1
mock==2.0.0
olefile==0.44
pbr==4.0.2
Pillow==4.3.0
psycopg2==2.7.4
pycodestyle==2.4.0
pylint==1.8.2
python-dateutil==2.6.1
python-decouple==3.1
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
urllib3==1.22
whitenoise==3.3.1
wrapt==1.10.11
